% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickdevicesservice_operations.R
\name{describe_device}
\alias{describe_device}
\title{Given a device ID, returns a DescribeDeviceResponse object describing the details of the device}
\usage{
describe_device(
  DeviceId = "string"
)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}
}
\description{
Given a device ID, returns a DescribeDeviceResponse object describing the details of the device.
}
\examples{

}
