% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickdevicesservice_operations.R
\name{get_device_methods}
\alias{get_device_methods}
\title{Given a device ID, returns the invokable methods associated with the device}
\usage{
get_device_methods(
  DeviceId = "string"
)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}
}
\description{
Given a device ID, returns the invokable methods associated with the device.
}
\examples{

}
