% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickdevicesservice_operations.R
\name{initiate_device_claim}
\alias{initiate_device_claim}
\title{Given a device ID, initiates a claim request for the associated device}
\usage{
initiate_device_claim(DeviceId)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}
}
\description{
Given a device ID, initiates a claim request for the associated device.
}
\details{
Claiming a device consists of initiating a claim, then publishing a device event, and finalizing the claim. For a device of type button, a device event can be published by simply clicking the device.
}
\section{Accepted Parameters}{
\preformatted{initiate_device_claim(
  DeviceId = "string"
)
}
}

