% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickdevicesservice_operations.R
\name{invoke_device_method}
\alias{invoke_device_method}
\title{Given a device ID, issues a request to invoke a named device method (with possible parameters)}
\usage{
invoke_device_method(DeviceId, DeviceMethod = NULL,
  DeviceMethodParameters = NULL)
}
\arguments{
\item{DeviceId}{The unique identifier of the device.}

\item{DeviceMethod}{The device method to invoke.}

\item{DeviceMethodParameters}{A JSON encoded string containing the device method request parameters.}
}
\description{
Given a device ID, issues a request to invoke a named device method (with possible parameters). See the "Example POST" code snippet below.
}
\examples{

}
