% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickdevicesservice_operations.R
\name{list_device_events}
\alias{list_device_events}
\title{Using a device ID, returns a DeviceEventsResponse object containing an array of events for the device}
\usage{
list_device_events(DeviceId, FromTimeStamp, ToTimeStamp, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{DeviceId}{The unique identifier of the device.}

\item{FromTimeStamp}{The start date for the device event query, in ISO8061 format. For example, 2018-03-28T15:45:12.880Z}

\item{ToTimeStamp}{The end date for the device event query, in ISO8061 format. For example, 2018-03-28T15:45:12.880Z}

\item{MaxResults}{The maximum number of results to return per request. If not set, a default value of 100 is used.}

\item{NextToken}{The token to retrieve the next set of results.}
}
\description{
Using a device ID, returns a DeviceEventsResponse object containing an array of events for the device.
}
\examples{

}
