% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickdevicesservice_operations.R
\name{list_devices}
\alias{list_devices}
\title{Lists the 1-Click compatible devices associated with your AWS account}
\usage{
list_devices(DeviceType = NULL, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{DeviceType}{The type of the device, such as "button".}

\item{MaxResults}{The maximum number of results to return per request. If not set, a default value of 100 is used.}

\item{NextToken}{The token to retrieve the next set of results.}
}
\description{
Lists the 1-Click compatible devices associated with your AWS account.
}
\examples{

}
