% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickdevicesservice_package.R
\docType{package}
\name{paws.iot1clickdevicesservice-package}
\alias{paws.iot1clickdevicesservice}
\alias{paws.iot1clickdevicesservice-package}
\title{paws.iot1clickdevicesservice: AWS IoT 1-Click Devices Service}
\description{
Stub description
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
