% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickdevicesservice_operations.R
\name{update_device_state}
\alias{update_device_state}
\title{Using a Boolean value (true or false), this operation enables or disables the device given a device ID}
\usage{
update_device_state(
  DeviceId = "string",
  Enabled = TRUE|FALSE
)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}

\item{Enabled}{If true, the device is enabled. If false, the device is disabled.}
}
\description{
Using a Boolean value (true or false), this operation enables or disables the device given a device ID.
}
\examples{

}
