% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickdevicesservice_operations.R
\name{unclaim_device}
\alias{unclaim_device}
\title{Disassociates a device from your AWS account using its device ID}
\usage{
unclaim_device(DeviceId)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}
}
\description{
Disassociates a device from your AWS account using its device ID.
}
\section{Accepted Parameters}{
\preformatted{unclaim_device(
  DeviceId = "string"
)
}
}

